Experiment: use webdriverio for tests
This builds on top of #52, but uses Webdriverio instead of Karma. It's been a long time since I've used it, and there have been significant changes (it can run unit tests via Mocha in the browser like Karma). This might be better in that it'll be one tool (but still two configurations and test runs) for both unit tests and e2e tests.