Support inline code and code blocks (#64)
Google Docs doesn’t have a native way to represent code, so we infer it based on text being in a monospaced font. This adds support both for inline code (with backticks in markdown) and for code blocks (indented in markdown).

For blocks, this uses indents instead of fences/triple-backticks, since indenting is the default for hast-util-to-mdast. It might be worth adding an option or revisiting this in the future (I personally like fences more, since you can set a language for code highlight in your Markdown processor later on).

I also wound up making some improvements to how overlapping inline formatting is handled (because of some edge cases around formatting inside code blocks), so the resulting markup is a little more streamlined. For example, this:

    This **is bold** **_and italic_**.

Is now:

    This **is bold _and italic_**.