Install correct Chromedriver version
Tests recently started failing in GitHub actions, and it's because the pre-installed version of Chrome and the version of Chromedriver specified in `package.json` fell out of sync. This installs the latest Chromedriver, but also adds some NPM configuration options to make this future-proof by installing the correct, matching versions of Chromedriver and Chromium itself, rather than the specific on in `package.json`.