Install correct Chromedriver version (#68)
Tests recently started failing in GitHub actions, and it's because the pre-installed version of Chrome and the version of Chromedriver specified in `package.json` fell out of sync. This installs the latest Chromedriver, but also adds an NPM configuration option to make this future-proof by detecting which version of Chrome is installed and downloading a matching Chromedriver version, rather than the specific version from `package.json`.