Add download button (#78)
There is now a “download” button alongside the “copy” button for the converted Markdown code. It saves the results to a file named `Converted Text.md`.

This also involved some complex additions to the tests in order to create temp directories that the browsers could download files to and be validated. Running tests now creates directories named `temp/chrome/`, `temp/firefox/`, and `temp/safari/` for each browser to work in. Unfortunately, there doesn’t seem to be an obvious way to set the download path for Safaridriver, so the test for this feature is skipped in that browser. :(

Co-authored-by: Rob Brackett <rob@robbrackett.com>