Specify minimum version of Node.js
This is essentially an alternative to #83. It specifies the minimum required version of Node.js in `package.json` (should have done that long ago!) and ignores various files for specifying the Node.js version in git, so we don't lock developers into a particular version (since we *intend* to eventually get project into NPM as a library and/or CLI tool and not just a web page, we don't want to lock in a specific version).