Allow scripts to work on Node.js 16
The proximate cause of #83 is that the development scripts (specifically `download-fixtures`) don't work on every Node.js version this package is meant to support. That's *ok* because they are only for use in development, but it's not great. This fixes things by including `undici` as a dev dependency so the scripts now work in all supported versions of Node.js.