Specify supported versions of Node.js (#100)
This supersedes #83. It specifies the minimum required version of Node.js in `package.json` (I should have done that long ago!) and ignores various files for specifying the Node.js version in git, so we don't lock developers into a particular version (since we *intend* to eventually get project into NPM as a library and/or CLI tool and not just a web page, we don't want to lock in a specific version).

It also adds `undici` as a dev dependency so that all the dev scripts actually work on all the versions of Node.js this package is meant to support, which was the original driving issue for #83.