Save Document Slice Clip data as fixtures (#111)
Since some upcoming features (#95) depend on the internal Google Docs "Document Slice Clip" data that may be present on the clipboard, we need fixtures for it! Luckily, recent changes in Chromium make this easier than it used to be (still not easy, though!).

The `download-fixtures` script now saves a `fixtures/<name>.copy.gdocsliceclip.json` file alongside the other fixture files. It’s kind of a long extension, but then the media type they use is long, too.