Make link anchors `<a>` tags and move to end
The `<a>` element is meant for exactly this use case, not just clickable links (that's why it's called "a" -- short for "anchor"). This also moves it to the end of headings instead of the start, since that's where the existing test expectation has it. As a result, the `internal-links` test now passes.