Remove bold/italic tag renaming
I could not find any clear examples of when/where this was needed in practice, and in the abstract, I think it's pretty debatable. Should the resulting markdown match the semantic info in the HTML or the styling in place? I think it's debatable; a `<strong>` could be restyled to not be bold but big, or colored, or underlined, or something else instead to show emphasis. It's not obvious that just because a `<strong>` is not bold it should not equate to `**text**` in markdown.