Clean up logging situation a bit
For errors and warnings, use `console.error|warn` directly so they don't get buried in hidden. For more debug/info level logs, continue using `debug` (and use the multiple argument form correctly, with formatting directives). I'm not super in love with `debug` in the browser, but it's already included by Micromark (used by Remark), so this is probably fine.