Fix deprecated ESLint rules
ESLint removed a bunch of formatting-related rules on the assumption that autoformatters like Prettier handle them and do a better job. For some cases, that's great; we can remove the rule and delegate to prettier. For others, we need the rule, which has moved out of ESLint core and into `@stylistic/eslint-plugin-js` (e.g. Prettier does not police the length of comments).