Pretty-format fixtures generated from GDocs
The `*.copy.html`, `*.export.html`, and `*.gdocsliceclip.json` fixture files generated from actual Google Docs are a huge pain to try and diff to understand what's changed and whether it affects us when Google Docs updates their formatting. I think it's important for the integrity of the tests that our fixtures parse exactly like an actual paste or export of a GDoc would, and so I've stored these fixtures exactly as we get them out of GDocs up until now.

Since the raw data has become such a pain to deal with, this adds routines for pretty-formatting the fixtures and then reversing that formatting when loading the fixtures for running tests. This should make diffs in git actually readable.

For JSON (GDocs Slice Clip format), this is pretty straightforward: we parse the `data` object from the Slice Clip envelope so it can be formatted as a JSON object just like the envelope itself, and then use a sorted JSON serializer so everything stays in a predictable order (so the file doesn't change on disk due to the order-less-ness of JS objects).

For HTML (copied and exported text), it's a bit messier. Whitespace between tags in HTML is significant, and most formatters will format something that parses at least a little bit differently from the original, and in a way where you can't reliably undo the formatting. That's fine for a lot of use cases, but not for tests in which we are trying to reliably match what happens with real Google Docs data. So this change includes a custom HTML formatter. It's not very robust and there are a lot of ways you can get it to produce invalid HTML, but a) we are unlikely to encounter those situations in our data and b) even when that happens, it still reverses just fine (so we might store some weird invalid HTML on disk, but it will be fine by the time it gets passed to a test). This could be made more reliable by involving a real HTML parser in the formatting, but that didn't seem worthwhile at this point. We can cross that bridge if and when we come to it.