Pretty-format fixtures generated from GDocs (#205)
Format our fixtures that are sourced from Google Docs so that they make for more readable diffs. This makes the fixtures as stored on disk a little different from what you’d actually get from copying/exporting a Google Doc, so this also involves tooling to *unformat* the fixtures when loading them in tests.

For Google Docs Slice Clip files (which are JSON), this is pretty simple and minimal: store the `data` field as an object instead of a JSON encoded string and sort the keys of objects (so the order stays stable — we arguably should *always* have been doing this; we’ve been lucky not to have problems so far).

For copied/exported HTML, this includes a pretty dumb custom formatter. Most HTML formatters make changes that can’t be perfectly reverted, and that’s a primary requirement here, since we want to maintain fidelity for tests. There are assuredly a lot of edge cases where this formatter is too dumb and could fail, but we aren’t likely to encounter those with actual Google Docs, and if we do, we can make improvements then.