Fix missing bold formatting in headings
When there's bold formatting in headings, we sometimes lose it because browsers normalize the styling and DOM structure when pasting content into a contenteditable node (which is how we currently capture a paste from Google Docs). Specifically, WebKit- and Chromium- based browsers would see that headings are already bold in the user agent stylesheet, and then normalize out all the bold formatting info in the actual DOM and CSS that gets pasted. This solves the issue by "resetting" the styling on headings inside the input area.

Fixes #113.