Keep heading IDs in docs w/ titles, empty headings
The previous heading finding algorithm for slice clips wound up with mismatches if a document included a title (which is a heading level 100 in the slice clip, but plain text in HTML) or headings that are empty (definitely a weird case; these don't show up at all in the HTML). This implements a new, more strict algorithm that also surfaces more info about the headings.

Fixes #215.