Add settings for heading IDs and code fences (#217)
Now that we have infrastructure for settings (from #213), I went ahead and added a couple other settings that have been long overdue:

- Fenced vs. indented code blocks.
- Heading IDs/links can be:
    - hidden (assuming you will autogenerate a heading ID based on the text later),
    - included as HTML `<a>` tags (for plain Markdown), or
    - included as a common extended syntax for element IDs (`{#id-here}`).

This makes the default for heading IDs "hidden", which is a bit of a change, but I have heard some folks feel the converstion got overly verbose and messy when we originally added heading links. And if you are explicitly specifying heading IDs, they may as well be short (even if random/meaningless).

Fixes #133.