Read code block data from slice clip
We previously support "code blocks" by looking for lines made entirely of monospaced text. This works well most of the time, but there are a few edge cases around formatting (see #93). We also weren't able to detect the language of a code block object. This changes adds support for reading the details out of the slice clip object if present, which allows us to supply the language info and solve the weird edge cases.

Fixes #93.