Read code snippet data from slice clip (#218)
We previously supported Google Docs code block/snippet objects by looking for sequences of lines made entirely of monospaced text. That worked well most of the time, but there are a few edge cases around formatting we couldn’t handle (see #93). We also weren't able to detect the language of a code snippet object.

This change adds support for reading the details out of the slice clip object if present, which allows us to supply the language info and solve the formatting edge cases.

Fixes #93.