Split up Slice Clip handling + HTML conversion
We previously let the browser handle pasting into our input area, and then ran a single converstion process on that HTML to make markdown. Since adding support for extra metadata that is only present in the "slice clip" format and not in HTML, this process has gotten more fragile: the slice clip tightly binds formatting info to specific indexes in the text, but our input area allows people to tweak the content after pasting, which now breaks the extra slice clip metadata.

This switches the architecture to convert in two steps:
1. On paste, merge the slice clip data we care about into the HTML from the clipboard, and drop the result into the input area.
2. On change of the input text, convert the HTML in the input to Markdown.

This allows people to teak the input, because in step 1 we converted the special slice clip data into a DOM-centric representation (e.g. bookmarks are now `<a>` elements, not arbitrary positions in the text).

This follows an approach that @trustedtomato first proposed in #96 and that I short-sightedly disagreed with at the time. I was previously concerned about the double-parsing involved here, but in hindsight, the robustness and features we get out of this approach are more than worth the efficiency cost. (There is also room for optimizing in the future.)

This is a very ugly first pass and needs cleanup.