Ensure fixtures use consistent asset URLs
The asset URLs in the Google Docs HTML (both copied and exported) include a unique-to-the-user key (presumably anonymized/hashed in some way). This meant each time we tried to refresh fixtures with images, we'd get a different result (since we are not logged in, we are a different user every time). This replaces the unique part of the asset URLs with a consistent key so our fixtures aren't pointlessly changing.

Obviates #284.