Don't rewrite fixture image if it has same hash
We generated a spurious fixture update today (#343) because a fixture image had the same visual hash but slightly different bytes. In this case, nothing really meaningful changed about the fixtures, and we shouldn't really have had an update. Instead, we should only write the image bytes to disk if the visual hash was different.