Minor vulnerable dependency updates (#359)
This should resolve minor security issues (just at dev time, so not a huge deal to begin with) except fora problem with esbuild. Unfortunately, it looks like we are stuck on vulnerable versions there because WebdriverIO and there is currently not compatible path to a secure version. That vulnerability only exists at test time, though, so it’s not a huge deal — just annoying.